Move everything to root
