+++ src/main/java/com/deltabook/DemoApplication.java
@@ -1,4 +1,4 @@
-package com.deltabook.demo;
+package com.deltabook;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
