+++ src/main/java/com/deltabook/model/Message.java
@@ -1,4 +1,4 @@
-package com.deltabook.demo.model;
+package com.deltabook.model;
 
 import javax.persistence.*;
 import java.sql.Timestamp;
