+++ src/main/java/com/deltabook/model/Roles.java
@@ -1,4 +1,4 @@
-package com.deltabook.demo.model;
+package com.deltabook.model;
 
 public enum Roles {
     ROLE_ADMIN,
