+++ src/main/java/com/deltabook/repositories/ContactRepository.java
@@ -1,7 +1,7 @@
-package com.deltabook.demo.repositories;
+package com.deltabook.repositories;
 
-import com.deltabook.demo.model.Contact;
-import com.deltabook.demo.model.User;
+import com.deltabook.model.Contact;
+import com.deltabook.model.User;
 import org.hibernate.validator.constraints.Range;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
