+++ src/main/java/com/deltabook/repositories/MessageRepository.java
@@ -1,7 +1,7 @@
-package com.deltabook.demo.repositories;
+package com.deltabook.repositories;
 
-import com.deltabook.demo.model.Message;
-import com.deltabook.demo.model.User;
+import com.deltabook.model.Message;
+import com.deltabook.model.User;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
