+++ src/main/resources/templates/registration.html
