+++ src/main/resources/templates/send_message.html
