+++ src/main/resources/templates/user_panel.html
